refactor(identity): share people query policy
