+++ src/backend/services/identity-resolution/src/api/persons.rs
@@ -37,7 +37,6 @@ use super::gate::require_admin;
 use super::listing::{self, CursorRejected};
 use super::resolution::PersonSummaryResponse;
 use crate::domain::person_card;
-use crate::domain::resolution::EXCLUDED_PERSON;
 use crate::infra::db::person_listing::{self, After, PersonListRow};
 use crate::infra::db::persons_repo;
 
@@ -84,7 +83,7 @@ pub async fn search_persons(
 
     let terms = search_terms(params.q.as_deref())?;
     let limit = listing::clamp_limit(params.limit, DEFAULT_LIMIT, MAX_LIMIT);
-    let (named, values) = partition_terms(&terms);
+    let (named, values) = listing::partition_person_terms(&terms);
 
     let query = terms.join(" ");
     let resume = resume_from(params.cursor.as_deref(), tenant, &query)?;
@@ -188,28 +187,6 @@ fn names_nobody(terms: &[String], named: &[Uuid], values: &[String]) -> bool {
     !terms.is_empty() && named.is_empty() && values.is_empty()
 }
 
-/// A term that parses as a UUID names a person id; everything else is matched
-/// against observed values.
-///
-/// Without this the one identifier an operator can copy off a card finds
-/// nothing, and a person the journal holds no attributes for — minted at first
-/// sign-in, before the resolver attaches the roster's name — cannot be found at
-/// all, since a value search has no value to match.
-fn partition_terms(terms: &[String]) -> (Vec<Uuid>, Vec<String>) {
-    let mut named = Vec::new();
-    let mut values = Vec::new();
-    for term in terms {
-        match Uuid::parse_str(term) {
-            // The excluded-person sentinel is not a person; naming it finds
-            // nobody rather than serving the row every exclusion appends to.
-            Ok(id) if id != EXCLUDED_PERSON => named.push(id),
-            Ok(_) => {}
-            Err(_) => values.push(term.clone()),
-        }
-    }
-    (named, values)
-}
-
 /// Split `q` into terms: non-empty, whitespace-separated, capped in count and
 /// total length. An absent or blank `q` is the whole roster, not an error.
 fn search_terms(q: Option<&str>) -> Result<Vec<String>, CanonicalError> {
@@ -232,6 +209,7 @@ mod tests {
     use std::error::Error;
 
     use super::*;
+    use crate::domain::resolution::EXCLUDED_PERSON;
 
     type R = Result<(), Box<dyn Error>>;
 
@@ -278,7 +256,7 @@ mod tests {
             "ann".to_owned(),
         ];
 
-        let (named, values) = partition_terms(&terms);
+        let (named, values) = listing::partition_person_terms(&terms);
 
         assert_eq!(named.len(), 1, "the id is a name, not a value to match");
         assert_eq!(values, vec!["ann".to_owned()]);
@@ -290,7 +268,7 @@ mod tests {
         // table — and it is still not a person the picker may offer.
         let terms = vec![EXCLUDED_PERSON.to_string()];
 
-        let (named, values) = partition_terms(&terms);
+        let (named, values) = listing::partition_person_terms(&terms);
 
         assert!(named.is_empty());
         assert!(values.is_empty(), "not matched as a value either");
